More updates and cleanup.
Switched to using pdflatex directly and turned on internal linking within
the pdf file.

Moved some things into moved_out.tex for the moment. They'll be reintegrated
back in if there is a natural place for them. (This includes all of the
figures, which will probably come back, but I had to move them out until
I make pdf versions of them to satisfy pdflatex.)